Seperate the API implementation of WebViewManager
